refactor: Remove number of visits variable in PlaceModel.
Increasing numVisits variable by 1 in a separate method when adding
visits could lead to issues where numVisits isn't updated to match
the actual size of the visits list, and vice versa. Instead, just
get the size of the visits list and display in the number of visits text.

Runs on API 28 emulator.