refactor: Get and modify visits directly using variable in DetailActivity.
Changed all code using getVisitGroups().getItems() to a use member variable
named visits, initially set to place.getVisits().
getVisitGroups().getItems() returns the exact same list as place.getVisits(),
but using two different methods for the same data is confusing.

More specifically:
getItems() returns list variable "items" which was set to place.getVisits() in
the ExpandableGroup constructor, so adding a visit to getItems() actually
adds a visit to place.getVisits() anyways.

Also, places list in MainActivity is cleared when the screen is rotated in MainActivity
or DetailActivity.

Runs on API 28 emulator and API 17 tablet.