feat: Insert place into database and display in RecyclerView.
The adapter's places list is set to the ViewModel's observed places list.
The insert method in MainActivity was changed to insert the place into the database.
The save and delete methods aren't implemented with the database yet.

The app incorrectly displays the empty list message after adding a place and closing
the app. The app also crashes when the added place is clicked and then the
save button is clicked. On API 28 emulator.