fix: insert place into database and display place in adapter.
The places list in MainActivity is set to the updated places list
from Observer's onChanged, which is the parameter in ViewModel's observe method.
Now empty list is displayed with empty list message before and after rotation, and
added place is displayed before and after rotation.

New places are being added to MainActivity member variablae places list but the places list is
being set to updated places list afterwards, which is redundant.

Remove and update to database have not been implemented.

Runs on API 28 emulator and API 17 tablet.