fix: Set date and time picker's initial date and time to clicked visit's date and time.
Refactored Visit to take a Calendar as the constructor so that the Visit's date and
time could be retrieved from Calendar fields and set for the DatePicker and TimePicker.
The Visit now has a Calendar as well as the date and time Strings.
Moved the code to convert the Calendar's date to a String time and String date into the
Visit class.

Runs on API 28 emulator and API 17 tablet.