refactor: Move all subtraction and addition to get visit position into DetailActivity.
Moving all position-related code into one file to make more maintainable and clearer.
The VisitGroup is at adapter position 0, and the visit list starts at adapter position 1, so
the first Visit is at adapter position 1.

Need to get a clicked visit's adapter position - 1 to get the visit's list position.
For example first visit is at adapter position 1 but visit list position 0.
The visit list position is used as a parameter to update, delete, and add visits.

Need to get visit's list position + 1 to get the visit's adapter position.
The adapter position is used to notify the adapter of changes to the visit list.

Runs on API 28 emulator and API 17 tablet.