Error loading photo in details screen.
Error is because Bitmap is fetched asynchronously with fetchPhoto().
The place model object selected in MainActivity is inserted before the Bitmap can
be loaded and set into the place model object. As a result the
place model object in details screen always has an empty Bitmap list and
no Bitmaps are displayed.