fix: Display place photo in details screen.
A selected place's Bitmap photo is fetched by using an asynchronous method fetchPhoto()
that waits for fetchPhotoResponse. The place is inserted before the method is finished, so
the ViewModel updates the place with the Bitmap once the method finishes.

To avoid a TransactionTooLargeException that crashes the app,
the Bitmap is given a maxiumum width and height and
then encoded to a Base64 String that is stored in the Room Database.
In DetailActivity, the Base64 String is retrieved and decoded to get the original Bitmap.
The Bitmap is displayed on the screen. Currently only the first photo is displayed.

Runs on API 28 emulator.