feat: User can add visits and see number of visits increase in manual place screen.
Number of visits updates if visit is added or deleted and resulting
visits are displayed in adapter when visit added, deleted, or updated.
Manual place is added to end of places list.
However, rotating the device makes the visits list empty and
sets the number of visits to 0 again.

Runs on API 28 emulator.