feat: Run a background service with notification.
Refresh places list service is an IntentService.
Currently the service just gets an input test String,
freezes background thread to simulate work,
and the service finishes in 10 seconds.
A notification is displayed on API 26 and higher while the Service is running,
and disappears once the Service is finished.

Runs on API 28 emulator.