feat: Get Place IDs from user's places list and log fetched place info.
New place info is logged when place can be found,
otherwise an error message is logged, ex. when place in list was added manually.
Notification is displayed with sound on older device when service runs,
but newer device doesn't play sound.
When app is closed by user, notification is still displayed on API 17 tablet.

Runs on API 28 emulator and API 17 tablet.