feat: Fetch place name and address if Place ID available, then them send to MainActivity.
When refresh menu item clicked, the IntentService to refresh the user's places list is
started and a notification is displayed. The IntentService is given all the Place IDs
of the places in the list.
The IntentService went through all Place IDs and fetched place name and address if
Place ID could be found.
Used BroadcastReceiver to receive place names and addresses.

Runs on API 28 emulator.