feat: Refresh places list with updated place names and addresses.
IntentService uses Place IDs to get up-to-date place names and addresses,
then new places are created with those Place IDs, names, and addresses.
The ViewModel is updated by replacing the user's original places with these new places.
A place isn't refreshed if it was added manually.

Fixed error of refreshed places info not being sent if last place in list was added manually.

Added comments that contain code to find number of bytes of String List.

If the place added by the user isn't able to fetch a photo by the time refresh menu item
is clicked, the new place will not have an image either.

Runs on API 28 emulator and API 17 tablet.