feat: Update widget when user saves a place.
Widget is updated becuase the name, address, and number of visits may change
when a user saves a place.

Additonally, moved widget so TextViews are not in the center, freeing up more space.
Changed AlertDialogs to specify the place or visit that is being deleted
to be more user-friendly.

Runs on API 28 emulator and API 17 tablet.