style: Remove unused code in comments.
Tested on API 17 tablet, but several errors occur.
1. The app takes a second to load the details screen when clicking a place.
2. If the user edits an EditText, ex. name or address, closes keyboard,
then tries to edit visits, ex. expands visits or adds a visit,
the screen will jump to the EditText. This might be because the EditText still
has focus.
3. Rotating the device several times causes an OutOfMemoryException with app crash.
The log mentions the Bitmap being decoded as the source of the error.
Added several visits before rotating, but not sure if that's related to the crash.