refactor: Move refresh places list functionality from service to main screen.
Using an IntentService to refresh places is unecessary since fetching places is
already done on a background thread. In addition, fetching place info in an IntentService and
then sending the info to a BroadcastReceiver through an Intent extra can cause a
TransactionTooLargeException if the info is too much data.

Tested on API 28 emulator but errors mentioned in previous commit still occur.
Refresh places functionality in MainActivity is okay.