feat: Long-press to drag and drop places.
User can long-press to reorder places in list. Clicking the place
opens up the correct details screen.

However, the app will load the original places list stored in database whenever
changes are observed, ex. the place is saved. When the user closes the app
and re-opens it the list will also be loaded as the original list.

Runs on API 28 emulator.