fix: Keep sort order when user drags and drops places.
Room Database sets sort position of each place before the place or manual place is added,
when places are refreshed, and after user interaction finishes in clearView method.

The sort position is stored in PlaceModel. This position is used by PlaceDao to order
all the places in getSortedPlaces. These places are observed by the ViewModel.

Now places sort order is kept even if user ex. saves a place or closes the app.

Runs on API 28 emulator and API 17 tablet.