feat: If edit text clicked, display handles and enable drag and drop.
If user clicks edit TextView, drag handles become visible and user can touch a
handle to drag a place. When the user clicks done the handles are gone and the
list positions of each place are saved in the database so
database can get all places in the correct order.

Runs on API 28 emulator and API 17 tablet.