refactor: Moved visit and visit group files to models package and adapter file to UI package.
Visit and VisitGroup are models. Placed them inside expandablegroup package to clearly
show the two are related to each other and the expanding RecyclerView functionality.
VisitGroupAdapter should be placed in UI package because it's part of UI. PlaceAdapter is also there and
adapters should be in the same package.

The source code structure is supposed to be packaged by feature. For example, UI and database are different features.

Runs on Pixel XL API 28 emulator and Asus API 17 tablet.
MainActivitySearchPlaceTest, MainActivityScreenTest, and AddManualPlaceTest pass.